Add multi PG version image Add multi patroni versions no defult for ARG DISTRO Add debug for ARG

+++ Dockerfile
@@ -1,19 +1,30 @@
-ARG DISTRO=alpine
-FROM postgres:17.6-${DISTRO}
+ARG DISTRO
+ARG PG_VERSION
+FROM postgres:${PG_VERSION}-${DISTRO}
 
-ENV PATRONI_VERSION=4.0.6-1
+ARG PATRONI_VERSION
+ENV PATRONI_VERSION=${PATRONI_VERSION}
+ARG DISTRO
 
 # Set the package manager and installation commands based on the distro
-RUN if [ "${DISTRO}" = "alpine" ]; then \
+RUN echo "DISTRO is: ${DISTRO}"  && echo "PATRONI_VERSION is: ${PATRONI_VERSION}" && \
+    if [ "${DISTRO}" = "alpine" ]; then \
       apk update \
       && apk add --no-cache \
-        patroni=${PATRONI_VERSION} \
-        py3-psycopg2; \
+        musl-locales \
+        python3 \
+        py3-pip \
+        py3-psycopg \
+        py3-psycopg-c \
+        py3-psycopg2 \
+        py3-psutil \
+      && pip install --no-cache-dir --break-system-packages \
+        "patroni[psycopg2,psycopg3,all]"==${PATRONI_VERSION} \
       && rm -rf /var/cache/apk/*; \
     else \
       apt update \
       && apt install -y --no-install-recommends \
-        patroni=${PATRONI_VERSION} \
+        patroni=${PATRONI_VERSION}-1 \
         python3-psycopg2 \
       && apt clean all \
       && rm -rf \
